Fix fatal error when running migrate
`php index.php migrate` results in a fatal error because the session is
undefined. From what I can tell, the profiler assumes the session is available,
which doesn't make sense for system actions like a database migration. To avoid
the issue, we just avoid the profiler when we're outside of a session.

https://github.com/anup-khanal-reisys/logger-codeigniter/issues/1